feat: add GitHub Actions workflow for deploying Quenta.ex 🚀
This commit introduces a new GitHub Actions workflow designed to automate the deployment of the Quenta.ex application. The goal is to ensure consistent deployments to the production environment whenever changes are pushed to the main branch. 

- Created a new workflow file `.github/workflows/deploy-quenta-ex.yml`
- Configured the workflow to trigger on pushes to the `main` branch affecting files in the `elixir` directory
- Added steps for checking out the code and setting up Flyctl for deployment
- Specified `flyctl deploy --remote-only ./elixir` to handle the deployment process

These changes are made to enhance the deployment process, allowing for smoother and automated updates of the application in the cloud environment.